requirements.txt
python-telegram-bot==21.9
httpx==0.28.1
beautifulsoup4==4.12.3
apscheduler==3.10.4
aiosqlite==0.19.0
python-dotenv==1.0.1
